# swagger-auth.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signout": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/user/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user's profile (JWT- required)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a user (owner or admin)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/remove/{id}": {
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove user (self or by admin) (JWT- required)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/all": {
      "get": {
        "operationId": "UserController_GetAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeUserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all users (admin only) (JWT- required)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/active/{id}": {
      "post": {
        "operationId": "UserController_active",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Toggle active status of a user (admin only) (JWT- required)",
        "tags": [
          "User"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "API documentation with JWT authentication",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Email użytkownika używany do logowania"
          },
          "password": {
            "type": "string",
            "example": "admin",
            "description": "Hasło użytkownika"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID użytkownika"
          },
          "roles": {
            "example": [
              "USER"
            ],
            "description": "Role użytkownika",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "accessToken": {
            "type": "string",
            "example": "jwt-access-token",
            "description": "Token dostępu JWT"
          },
          "refreshToken": {
            "type": "string",
            "example": "jwt-refresh-token",
            "description": "Token odświeżający JWT"
          }
        },
        "required": [
          "id",
          "roles",
          "accessToken",
          "refreshToken"
        ]
      },
      "Item": {
        "type": "object",
        "properties": {}
      },
      "Bid": {
        "type": "object",
        "properties": {}
      },
      "Review": {
        "type": "object",
        "properties": {}
      },
      "SafeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID użytkownika"
          },
          "name": {
            "type": "string",
            "example": "Jan Kowalski",
            "description": "Imię i nazwisko użytkownika"
          },
          "email": {
            "type": "string",
            "example": "jan.kowalski@example.com",
            "description": "Email użytkownika"
          },
          "roles": {
            "example": [
              "USER"
            ],
            "description": "Role użytkownika",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Czy użytkownik jest aktywny"
          },
          "verified": {
            "type": "boolean",
            "example": false,
            "description": "Czy użytkownik jest zweryfikowany"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg",
            "description": "URL avatara użytkownika"
          },
          "createdAt": {
            "type": "string",
            "description": "Data utworzenia konta",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data ostatniej aktualizacji",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "description": "Data ostatniego logowania",
            "format": "date-time"
          },
          "favorites": {
            "description": "Ulubione przedmioty użytkownika",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "bids": {
            "description": "Oferty użytkownika",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bid"
            }
          },
          "ownedItems": {
            "description": "Przedmioty wystawione przez użytkownika",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "reviews": {
            "description": "Recenzje wystawione przez użytkownika",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          }
        },
        "required": [
          "id",
          "email",
          "roles",
          "isActive",
          "verified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "name",
          "email"
        ]
      }
    }
  }
}